% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fec_vars}
\alias{fec_vars}
\title{FEC variables}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 111 rows and 9 columns.}
\usage{
fec_vars
}
\description{
FEC variables
}
\keyword{datasets}
